Separate Tornado handlers into their own class
