+++ voila/treehandler.py
@@ -6,14 +6,9 @@
 #                                                                           #
 # The full license is in the file LICENSE, distributed with this software.  #
 #############################################################################
-import os
-
 from jupyter_server.utils import url_escape, url_path_join
-from tornado import web
 
 from .handler import BaseVoilaHandler
-from .utils import get_server_root_dir
-from nbclient.util import ensure_async
 
 
 class VoilaTreeHandler(BaseVoilaHandler):
@@ -50,55 +45,3 @@ class VoilaTreeHandler(BaseVoilaHandler):
             return page_title + "/"
         else:
             return "Voilà Home"
-
-    @web.authenticated
-    async def get(self, path=""):
-        cm = self.contents_manager
-        dir_exists = await ensure_async(cm.dir_exists(path=path))
-        file_exists = await ensure_async(cm.file_exists(path))
-        if dir_exists:
-            is_hidden = await ensure_async(cm.is_hidden(path))
-            if is_hidden and not cm.allow_hidden:
-                self.log.info("Refusing to serve hidden directory, via 404 Error")
-                raise web.HTTPError(404)
-            breadcrumbs = self.generate_breadcrumbs(path)
-            page_title = self.generate_page_title(path)
-            contents = await ensure_async(cm.get(path))
-
-            def allowed_content(content):
-                if content["type"] in ["directory", "notebook"]:
-                    return True
-                __, ext = os.path.splitext(content.get("path"))
-                return ext in self.allowed_extensions
-
-            contents["content"] = sorted(contents["content"], key=lambda i: i["name"])
-            contents["content"] = filter(allowed_content, contents["content"])
-
-            self.write(
-                self.render_template(
-                    "tree.html",
-                    frontend="voila",
-                    main_js="voila.js",
-                    page_title=page_title,
-                    notebook_path=path,
-                    breadcrumbs=breadcrumbs,
-                    contents=contents,
-                    terminals_available=False,
-                    server_root=get_server_root_dir(self.settings),
-                    query=self.request.query,
-                )
-            )
-        elif file_exists:
-            # it's not a directory, we have redirecting to do
-            model = await ensure_async(cm.get(path, content=False))
-            # redirect to /api/notebooks if it's a notebook, otherwise /api/files
-            service = "notebooks" if model["type"] == "notebook" else "files"
-            url = url_path_join(
-                self.base_url,
-                service,
-                url_escape(path),
-            )
-            self.log.debug("Redirecting %s to %s", self.request.path, url)
-            self.redirect(url)
-        else:
-            raise web.HTTPError(404)
